fix: address 3 deferred CodeRabbit Review 1 items (#20, #23, #28)
- #20: YahavDatePicker .pmu-* hardcoded CSS moved to ScraperConfig
  (yearsGridPrefix, daysGridPrefix, monthsGridPrefix, daysGridCheck)
- #23: MizrahiFixtures mockLocator delegates to buildMockLocator (DRY)
- #28: VisaCalIntegration shared createIframeLocatorMock in VisaCalFixtures

All files lint-checked after each edit. Zero back-and-forth.
